Revert "Merge branch 'devMaster' into master"
This reverts commit ecd82a95f178312466725325d85a684ce6e49072, reversing
changes made to 263329fa0fc087618fe436fcead4530670a08e89.
