+++ src/components/App/index.js
@@ -12,9 +12,9 @@ import HomePage from '../Home';
 import AccountPage from '../Account';
 import AdminPage from '../Admin';
 import AgentPage from '../Agent';
-import ClientPage from '../Client';
 
 import * as ROUTES from '../../constants/routes';
+import * as ROLES from '../../constants/roles';
 import { withAuthentication } from '../Session';
 const App = () => (
   <Router>
@@ -26,8 +26,6 @@ const App = () => (
       <Route path={ROUTES.SIGN_IN} component={SignInPage} />
       <Route path={ROUTES.PASSWORD_FORGET} component={PasswordForgetPage}/>
       <Route path={ROUTES.HOME} component={HomePage} />
-      <Route path={ROUTES.AGENT} component={AgentPage} />
-      <Route path={ROUTES.CLIENT} component={ClientPage} />
       <Route path={ROUTES.ACCOUNT} component={AccountPage} />
       <Route path={ROUTES.ADMIN} component={AdminPage} />
       <Route path={ROUTES.PASSWORD_CHANGE} component={PasswordChangePage} />
