+++ src/components/PasswordChange/index.js
@@ -1,5 +1,5 @@
 import React, {Component} from 'react';
-import {Link} from 'react-router-dom';
+import {Link, withRouter } from 'react-router-dom';
 
 import {withFirebase} from '../Firebase';
 import * as ROUTES from '../../constants/routes';
