+++ src/constants/routes.js
@@ -5,9 +5,6 @@ export const HOME = '/home';
 export const ACCOUNT = '/account';
 export const ADMIN = '/admin';
 export const ADMIN_DETAILS = '/admin/:id';
-export const AGENT = '/agent';
-export const AGENT_DETAILS = '/agent/:id';
-export const CLIENT = '/client';
-export const CLIENT_DETAILS = '/client/:id';
+export const USER_PAGE = '/mypage';
 export const PASSWORD_FORGET = '/pw-forget';
 export const PASSWORD_CHANGE = '/pw-change';
